Re-organizing header files
I'm making a solid attempt to clean up the header files. I would like to
move as much as possible out of the sensor specific headers and into the
common file. Furhter, I would like to separate each of the types into
separate headers based on algorithm section (to which the type applies).

This should greatly add to code clarity as well as provide easier
modification in the future.