Rebuild (#3)
* Recplaced all globals with pointers

This switch should add code clairity for the C Emulator implementation.
Note that the IMU firmware will still be using globals, however the
function calls will be sending the address. Since there are a large
number of functions, making this switch should make each functions role
more explicit.

* Converting globals to pointers

* Compiled

I have no idea if it works, but it compiles...

* Beginning Refactoring Part 2

I weaved in pointers rather than ambiguous global references. This
should help code clarity.

Part 2:
Get it to work.
I'll make a single sweep through the code and clean up any
inconsistancies I see and ensure proper code format. Then I'll run a few
tests.
This next section should see some changes to the header files as well as
the #defines. The goal is to get the gains to be set by g_control rather
than at compile. This will allow for the dynamic modification of the
algorithms.

* cleaning up main

* Re-organizing header files

I'm making a solid attempt to clean up the header files. I would like to
move as much as possible out of the sensor specific headers and into the
common file. Furhter, I would like to separate each of the types into
separate headers based on algorithm section (to which the type applies).

This should greatly add to code clarity as well as provide easier
modification in the future.

* Modified Init functions

Compartmentalizing initializations routines.

* Re-organizing headers

knee deep in the headers. There are some new headers, renamed headers,
new function files, etc.

* Debugging

I've re-organize a lot of the header files.
I've still a bit to go, but I've got quite a bit done.
I've debugged a little. It compiles but I assume it will not work out of
the box....

* Header files are near complete

The header files are nearing maturity.
Beginning work on the functions themselves.

* Nearly there

Finished adding a few more fixes. Need to debug and preform some general
cleanup.

* Getting closer

There seems to be a bug in the gapa routine.
Results don't match matlab.
DCM comparisons are very close. Could simply be roundoff.

* Implementing Changes from Arduino Merge

Added the updates to the logging functions to the C implementation to
ensure future development.

* Added logging functionality to C implementation

This is along the lines of general cleanup.

* Added start/stop detection

Start/stop based on gyro magnitude is included.

* added default z scaling parameters

default values, set in the gapa header file, should now produce results
similar to the matlab code.